+++ src/main/kotlin/shop/itbug/fluttercheckversionx/document/MarkdownRender.kt
@@ -198,12 +198,18 @@ fun MarkdownNode.toHtml(project: Project): String {
             MarkdownElementTypes.INLINE_LINK -> {
                 val label = node.child(MarkdownElementTypes.LINK_TEXT)?.toHtml(project)
                 val destination = node.child(MarkdownElementTypes.LINK_DESTINATION)?.text
-                if (label != null && destination != null) {
-                    val atag = HtmlChunk.tag("a").attr("href", destination).addText(label).toString()
-                    sb.append(atag)
+                val videoHtmlText = generateFlutterMp4Preview(destination ?: "")
+                if (videoHtmlText != null) {
+                    sb.append(videoHtmlText)
                 } else {
-                    sb.append(node.text)
+                    if (label != null && destination != null) {
+                        val atag = HtmlChunk.tag("a").attr("href", destination).addText(label).toString()
+                        sb.append(atag)
+                    } else {
+                        sb.append(node.text)
+                    }
                 }
+
             }
 
             MarkdownTokenTypes.TEXT,
@@ -306,7 +312,17 @@ fun MarkdownNode.toHtml(project: Project): String {
             }
         }
     }
-    return sb.toString().trimEnd()
+    val fullHtml = sb.toString().trimEnd()
+    return fullHtml
+}
+
+///解析mp4链接
+fun generateFlutterMp4Preview(url: String): String? {
+    if (url.endsWith(".mp4") && url.startsWith("https://flutter.github.io")) {
+        val html = HtmlChunk.div().child(HtmlChunk.tag("a").attr("href", url).addText(url)).toString()
+        return html
+    }
+    return null
 }
 
 /**